Refactor api
